Рефакторинг недельного расписания Оно настолько сложное, что пришлось ещё один модуль выводить
